+++ lib/docking.dart
@@ -3,4 +3,5 @@ library docking;
 export 'src/layout/docking_layout.dart';
 export 'src/on_item_selection.dart';
 export 'src/on_item_close.dart';
+export 'src/docking_buttons_builder.dart';
 export 'src/docking.dart';